refactor(stats): split ProgressStats into smaller components and optimize performance by removing unnecessary memo
